update readme add support for workflow-dispatch

+++ src/Extract-PBIXMetadata.ps1
@@ -12,10 +12,12 @@ Import-Module -Name MicrosoftPowerBIMgmt
 
 $git_event_before = $env:GIT_EVENT_BEFORE
 $git_event_after = $env:GIT_EVENT_AFTER
+$git_trigger_Event = $env:GIT_TRIGGER_EVENT
 
 $workspace_id = $env:PBI_PREMIUM_WORKSPACE_ID
 
-if($env:PBI_TENANT_ID -and $env:PBI_CLIENT_ID -and $env:PBI_CLIENT_SECRET) {
+
+if ($env:PBI_TENANT_ID -and $env:PBI_CLIENT_ID -and $env:PBI_CLIENT_SECRET) {
 	Write-Information "Using Service Principal authentication!"
 	$tenant_id = $env:PBI_TENANT_ID
 	$client_id = $env:PBI_CLIENT_ID
@@ -41,17 +43,22 @@ else {
 
 $workspace = Get-PowerBIWorkspace -Id $workspace_id
 
-# get the changed .pbix files in the current push
-$changed_files = Join-Path $root_path "_tmp_changed_files.txt"
-$x = Start-Process "git" -ArgumentList @("diff", "--name-only", $git_event_before, $git_event_after, "--diff-filter=ACM", """*.pbix""") -Wait -PassThru -NoNewWindow -RedirectStandardOutput $changed_files
-#$x = Start-Process "git" -ArgumentList @("diff", "--name-only", "HEAD~2", """*.pbix""") -Wait -PassThru -NoNewWindow -RedirectStandardOutput $changed_files
-$pbix_files = Get-Content -Path $changed_files | ForEach-Object { Join-Path $root_path $_ | Get-Item}
-Remove-Item $changed_files
+if ($get_trigger_event -eq "push") {
+	# get the changed .pbix files in the current push
+	$changed_files = Join-Path $root_path "_tmp_changed_files.txt"
+	$x = Start-Process "git" -ArgumentList @("diff", "--name-only", $git_event_before, $git_event_after, "--diff-filter=ACM", """*.pbix""") -Wait -PassThru -NoNewWindow -RedirectStandardOutput $changed_files
+	#$x = Start-Process "git" -ArgumentList @("diff", "--name-only", "HEAD~2", """*.pbix""") -Wait -PassThru -NoNewWindow -RedirectStandardOutput $changed_files
+	$pbix_files = Get-Content -Path $changed_files | ForEach-Object { Join-Path $root_path $_ | Get-Item }
+	Remove-Item $changed_files
+}
+elseif ($get_trigger_event -eq "workflow_dispatch") {
+	$pbix_files = Get-ChildItem -Recurse -Filter "*.pbix"
+}
+
 
 Write-Information $pbix_files
 
-foreach($pbix_file in $pbix_files)
-{
+foreach ($pbix_file in $pbix_files) {
 	$report = $null
 	$dataset = $null
 	try {
@@ -59,8 +66,7 @@ foreach($pbix_file in $pbix_files)
 
 		Write-Information "Checking PBIX file contains a datamodel ..."
 		$zip_entries = [IO.Compression.ZipFile]::OpenRead($pbix_file.FullName).Entries.Name; 
-		if("DataModel" -notin $zip_entries)
-		{
+		if ("DataModel" -notin $zip_entries) {
 			Write-Information "No datamodel found in $($pbix_file.Name) - skipping further processing of this file!"
 			continue
 		}
@@ -75,7 +81,7 @@ foreach($pbix_file in $pbix_files)
 		Write-Information "    Done!"
 
 		Write-Information "Getting PowerBI dataset ..."
-		$dataset = Get-PowerBIDataset -WorkspaceId $workspace.Id | Where-Object { $_.Name -eq $temp_name}
+		$dataset = Get-PowerBIDataset -WorkspaceId $workspace.Id | Where-Object { $_.Name -eq $temp_name }
 		$connection_string = "powerbi://api.powerbi.com/v1.0/myorg/$($workspace.Name);initial catalog=$($dataset.Name)"
 
 		$executable = Join-Path $root_path TabularEditor.exe
@@ -89,8 +95,7 @@ foreach($pbix_file in $pbix_files)
 		Write-Information "$executable $params"
 		$p = Start-Process -FilePath $executable -Wait -NoNewWindow -PassThru -ArgumentList $params
 
-		if($p.ExitCode -ne 0)
-		{
+		if ($p.ExitCode -ne 0) {
 			Write-Error "Failed to extract .bim file from $($dataset.WebUrl)!"
 		}
 	}
@@ -98,13 +103,11 @@ foreach($pbix_file in $pbix_files)
 		
 	}
 	finally {
-		if($report -ne $null)
-		{
+		if ($report -ne $null) {
 			Write-Information "Removing temporary PowerBI report ..."
 			Remove-PowerBIReport -WorkspaceId $workspace.Id -Id $report.Id
 		}
-		if($dataset -ne $null)
-		{
+		if ($dataset -ne $null) {
 			Write-Information "Removing temporary PowerBI dataset ..."
 			Invoke-PowerBIRestMethod -Url "https://api.powerbi.com/v1.0/myorg/groups/$($workspace.Id)/datasets/$($dataset.Id)" -Method Delete
 		}
